Move Projects.js to new SearchProjects folder
